Move visualization back to the parent directory
