+++ docs/source/visualization.rst
@@ -1,13 +1,13 @@
 Visualization tool
 ==================
 
-:mod:`graphix.flow.visualization` module
+:mod:`graphix.visualization` module
 +++++++++++++++++++++++++++++++++++++++++
 
 This module provides functions to visualize the resource state of MBQC pattern.
 If flow or gflow exist, the tool take them into account and show the information flow as directed edges.
 
-.. currentmodule:: graphix.flow.visualization
+.. currentmodule:: graphix.visualization
 
 .. autoclass:: GraphVisualizer
     :members:
